Complete mysql seeding script
